pyproject.toml
[tool.poetry]
name = "hospital-equipment-maintenance"
version = "0.1.0"
description = "Hospital Equipment Maintenance Management and Reminder System"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
pydantic = "^2.5.0"
pandas = "^2.1.0"
schedule = "^1.2.0"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"  # For scheduling background tasks
email-validator = "^2.1.0"
flask-wtf = "^1.2.1"  # For form validation

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
